Move chat log logic from CardAbility to CardAbilityStep
